create Kafka.py
This is a simple script which abstracts some of the "connecting to Kafka code" and this lets you create producers and create infinitely running consumers. I wrote this script for one of my projects in the initial stage of development but switched to a different approach for cutting deployment cost.